meta_srv::discovery

Module utils

Traits

LastActiveTs

Trait for types that have a last active timestamp.

Functions

accept_ingest_workload

Determines if a datanode is capable of accepting ingest workloads. Returns true if the datanode’s workload types include ingest capability, or if the node is not of type [NodeWorkloads::Datanode].

alive_datanode

Returns the alive datanode peer.

alive_datanodes

Returns the alive datanodes.

alive_flownodes

Returns the alive flownodes.

alive_frontends

Returns the alive frontends.

build_active_filter

Builds a filter closure that checks whether a LastActiveTs item is still within the specified active duration, relative to the current time provided by the given [SystemTimer].

find_datanode_lease_value

Returns the lease value of the given datanode id, if the datanode is not found, returns None.