Trait meta_srv::election::Election

pub trait Election: Send + Sync {
    type Leader;

    // Required methods
    fn is_leader(&self) -> bool;
    fn in_infancy(&self) -> bool;
    fn campaign<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn leader<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Self::Leader>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn resign<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe_leader_change(&self) -> Receiver<LeaderChangeMessage>;
}

Required Associated Types

type Leader

Required Methods

fn is_leader(&self) -> bool

Returns true if current node is the leader.

fn in_infancy(&self) -> bool

When a new leader is born, it may need some initialization operations (asynchronous), this method tells us when these initialization operations can be performed.

note: a new leader will only return true on the first call.

fn campaign<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Campaign waits to acquire leadership in an election.

Multiple sessions can participate in the election, but only one can be the leader at a time.

fn leader<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Self::Leader>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the leader value for the current election.

fn resign<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Releases election leadership so other campaigners may acquire leadership on the election.

fn subscribe_leader_change(&self) -> Receiver<LeaderChangeMessage>