meta_srv

Module gc

Modules

candidate 🔒

ctx 🔒

handler 🔒

options 🔒

procedure 🔒

scheduler 🔒

tracker 🔒

Structs

BatchGcProcedure

Procedure to perform get file refs then batch GC for multiple regions, should only be used by admin function for triggering manual gc, as it holds locks for too long and for all regions during the procedure.

GcSchedulerOptions

Configuration for GC operations.

Type Aliases

Peer2Regions 🔒

Region2Peers