meta_srv::gc

Module procedure

Structs

BatchGcData

BatchGcProcedure

Procedure to perform get file refs then batch GC for multiple regions, should only be used by admin function for triggering manual gc, as it holds locks for too long and for all regions during the procedure.

GcRegionData

GcRegionProcedure

TODO(discord9): another procedure which do both get file refs and gc regions.

Enums

State

Functions

send_gc_regions 🔒

Helper function to send GcRegions instruction and wait for reply.

send_get_file_refs 🔒

Helper function to send GetFileRefs instruction and wait for reply.