meta_srv::procedure::repartition

Module group

Modules

apply_staging_manifest ğŸ”’

enter_staging_region ğŸ”’

remap_manifest ğŸ”’

repartition_end ğŸ”’

repartition_start ğŸ”’

sync_region ğŸ”’

update_metadata ğŸ”’

utils ğŸ”’

Structs

Context

GroupPrepareResult

The result of the group preparation phase, containing validated region routes.

Metrics

PersistentContext

RepartitionGroupData ğŸ”’

RepartitionGroupDataOwned ğŸ”’

RepartitionGroupProcedure

VolatileContext

Traits

State ğŸ”’

Functions

region_routes

Returns the region routes of the given table route value.

Type Aliases

GroupId