meta_srv::procedure::repartition::group

Module repartition_start

Structs

RepartitionStart

Functions

ensure_region_route_expr_match 🔒

Ensures that the partition expression of the region route matches the partition expression of the region descriptor.