meta_srv::procedure::wal_prune

Module utils

Constants

DEFAULT_PARTITION 🔒

The default partition.

DELETE_RECORDS_TIMEOUT 🔒

The default timeout for deleting records.

Functions

delete_records 🔒

Deletes the records for the given topic.

find_pruneable_entry_id_for_topic 🔒

Finds the prunable entry id for the topic.

get_offsets_for_topic 🔒

Returns the earliest and latest offsets for the given topic.

get_partition_client 🔒

Returns a partition client for the given topic.

missing_region_ids 🔒

should_trigger_prune 🔒

Determines whether pruning should be triggered based on the current pruned entry id and the prunable entry id. Returns true if:

update_pruned_entry_id 🔒

Updates the pruned entry id for the given topic.