Module meta_srv::pubsub

Modules

publish 🔒

subscribe_manager 🔒

subscriber 🔒

Structs

AddSubRequest

DefaultPublish

The default implementation of Publish

DefaultSubscribeManager

Subscriber

UnSubRequest

Enums

Message

Topic

Subscribed topic type, determined by the ability of meta.

Traits

Publish

This trait provides a send_msg method that can be used by other modules of meta to publish Message.

SubscribeManager

SubscribeQuery

Transport

This trait defines how messages are delivered from meta to the subscriber.

Type Aliases

PublishRef

SubscribeManagerRef

SubscriberRef