meta_srv::region

Module flush_trigger

Structs

RegionFlushTicker 🔒

RegionFlushTicker is used to trigger region flush trigger periodically.

RegionFlushTrigger

RegionFlushTrigger is used to ensure that the estimated WAL replay size stays below a certain threshold by triggering a region flush when the estimated WAL replay size exceeds that threshold. This helps improve datanode startup speed and reduce the overall startup time.

Enums

Event 🔒

Event represents various types of events that can be processed by the region flush ticker.

Constants

RECENT_DURATION 🔒

The duration of the recent period.

TICKER_INTERVAL 🔒

The interval of the region flush ticker.

Functions

filter_regions_by_replay_size 🔒

Filter regions based on the estimated replay size.

group_regions_by_leader 🔒

Group regions by leader.

is_recent 🔒

Check if the timestamp is recent.

should_persist_region_checkpoint 🔒

Determines whether a region checkpoint should be persisted based on current and persisted state.

Type Aliases

RegionFlushTickerRef 🔒