meta_srv::failure_detector

Function phi

fn phi(time_diff: i64, mean: f64, std_deviation: f64) -> f64

Calculation of phi, derived from the Cumulative distribution function for N(mean, stdDeviation) normal distribution, given by 1.0 / (1.0 + math.exp(-y * (1.5976 + 0.070566 * y * y))) where y = (x - mean) / standard_deviation This is an approximation defined in β Mathematics Handbook (Logistic approximation). Error is 0.00014 at +- 3.16 The calculated value is equivalent to -log10(1 - CDF(y))

Usually phi = 1 means likeliness that we will make a mistake is about 10%. The likeliness is about 1% with phi = 2, 0.1% with phi = 3 and so on.