meta_srv::procedure
Module region_migration

Modules

close_downgraded_region 🔒

downgrade_leader_region 🔒

manager 🔒

migration_abort 🔒

migration_end 🔒

migration_start 🔒

open_candidate_region 🔒

update_metadata 🔒

upgrade_candidate_region 🔒

Structs

Context

The context of procedure execution.

DefaultContextFactory

Default implementation.

PersistentContext

It’s shared in each step and available even after recovering.

RegionMigrationData

Persistent data of RegionMigrationProcedure.

RegionMigrationDataOwned

Persistent data of RegionMigrationProcedure.

RegionMigrationProcedure 🔒

RegionMigrationProcedureTask

RegionMigrationProcedureTracker

VolatileContext

It’s shared in each step and available in executing (including retrying).

Traits

ContextFactory

Used to generate new Context.

State 🔒

Functions

default_timeout 🔒

Type Aliases

RegionMigrationManagerRef