meta_srv::procedure::wal_prune

Module manager

Structs

WalPruneManager 🔒

WalPruneManager manages all remote WAL related tasks in metasrv.

WalPruneProcedureGuard

WalPruneProcedureGuard is a guard for WalPruneProcedure. It is used to track the running WalPruneProcedures. When the guard is dropped, it will remove the topic name from the running procedures and release the semaphore.

WalPruneProcedureTracker

Tracks running WalPruneProcedures and the resources they hold. A WalPruneProcedure is holding a semaphore permit to limit the number of concurrent procedures.

WalPruneTicker 🔒

WalPruneTicker is a ticker that periodically sends Events to the WalPruneManager. It is used to trigger the WalPruneManager to submit WalPruneProcedures.

Enums

Event

Event is used to notify the WalPruneManager to do some work.

Type Aliases

WalPruneTickerRef