meta_srv
Module pubsub

Modules

publish 🔒

subscribe_manager 🔒

subscriber 🔒

Structs

DefaultPublisher

The default implementation of Publisher

DefaultSubscribeManager

SubscribeRequest

Subscriber

UnsubscribeRequest

Enums

Message

Topic

Subscribed topic type, determined by the ability of meta.

Traits

Publisher

This trait provides a publish method that can be used by other modules of meta to publish Message.

SubscriptionManager

SubscriptionQuery

Transport

This trait defines how messages are delivered from meta to the subscriber.

Type Aliases

PublisherRef

SubscriberRef

SubscriptionManagerRef