meta_srv::pubsub::publish

Trait Publisher

pub trait Publisher: Send + Sync {
    // Required method
    fn publish<'life0, 'async_trait>(
        &'life0 self,
        message: Message,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

This trait provides a publish method that can be used by other modules of meta to publish Message.

Required Methods

fn publish<'life0, 'async_trait>( &'life0 self, message: Message, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,